requirements.txt
asgiref==3.8.1
asttokens==2.4.1
cfgv==3.4.0
decorator==5.1.1
distlib==0.3.8
Django==5.0.4
django-extensions==3.2.3
djangorestframework==3.15.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.4
flake8==7.0.0
gunicorn==21.2.0
identify==2.5.35
ipython==8.23.0
isort==5.13.2
jedi==0.19.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.0
pre-commit==3.7.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.2
typing_extensions==4.11.0
virtualenv==20.25.1
wcwidth==0.2.13
